SOC-3139 | refactor: extract code that generates mobile application links to common class
