+++ extensions/wikia/Email/Email.setup.php
@@ -31,6 +31,7 @@ $wgAutoloadClasses['Email\Helper'] =  $dir . 'EmailHelper.class.php';
 $wgAutoloadClasses['Email\ImageHelper'] =  $dir . 'EmailImageHelper.class.php';
 $wgAutoloadClasses['Email\ControllerException'] =  $dir . 'EmailExceptions.class.php';
 $wgAutoloadClasses['Email\EmailMobileBadges'] =  $dir . 'EmailMobileBadges.class.php';
+$wgAutoloadClasses['Email\MobileApplicationsLinksGenerator'] =  $dir . 'MobileApplicationsLinksGenerator.class.php';
 $wgAutoloadClasses['Email\Fatal'] =  $dir . 'EmailExceptions.class.php';
 $wgAutoloadClasses['Email\Check'] =  $dir . 'EmailExceptions.class.php';
 $wgAutoloadClasses['Email\Controller\ForgotPasswordController'] =  $dir . 'Controller/ForgotPasswordController.class.php';
