+++ extensions/wikia/Email/EmailController.class.php
@@ -2,7 +2,6 @@
 
 namespace Email;
 
-use Http;
 use TijsVerkoyen\CssToInlineStyles\CssToInlineStyles;
 use Wikia\Logger\WikiaLogger;
 
@@ -12,12 +11,6 @@ abstract class EmailController extends \WikiaController {
 	const TRACKED_LANGUAGES = [ 'EN', 'PL', 'DE', 'ES', 'FR', 'IT', 'JA', 'NL', 'PT', 'RU', 'ZH' ];
 
 	const AVATAR_SIZE = 40;
-
-	const ANDROID_PLATFORM = 'android';
-
-	const IOS_PLATFORM = 'ios';
-
-	const MOBILE_APPLICATIONS_LINKS_EVICTION_TIME = 24 * 60 * 60; // 24h
 
 	/** CSS used for the main content section of each email. Used by getContent()
 	 * and intended to be overridden by child classes. */
@@ -413,7 +406,8 @@ abstract class EmailController extends \WikiaController {
 	}
 
 	private function generateMobileApplicationsBadges() {
-		$mobileApplicationsLinks = $this->generateMobileApplicationsLinks();
+		$linksGenerator = new MobileApplicationsLinksGenerator();
+		$mobileApplicationsLinks = $linksGenerator->generate();
 		$hasMobileApplicationBadges = count( $mobileApplicationsLinks ) > 0;
 		$badges = null;
 
@@ -429,90 +423,6 @@ abstract class EmailController extends \WikiaController {
 		}
 
 		return $badges;
-	}
-
-	private function generateMobileApplicationsLinks() {
-		global $wgMemc;
-
-		$mobileApplicationsLinks =
-			$wgMemc->get( $this->createCacheKeyForMobileApplicationsLinks() );
-
-		if ( is_bool( $mobileApplicationsLinks ) ) {
-			$mobileApplicationsLinks = [];
-			$response = $this->fetchMobileApplicationsDetails();
-
-			if ( $response && $this->applicationsExistIn( $response ) ) {
-				$mobileApplications = json_decode( $response, true );
-				$mobileApplicationsLinks = $this->traverseThrough( $mobileApplications );
-			} else {
-				$mobileApplicationsLinks[self::ANDROID_PLATFORM] =
-					'https://play.google.com/store/apps/developer?id=Wikia,+Inc.';
-				$mobileApplicationsLinks[self::IOS_PLATFORM] =
-					'https://itunes.apple.com/us/developer/wikia-inc./id422467077';
-			}
-
-			// Event if response is not valid (for example there are some difficulties with mobile applications service)
-			// result will be cached to prevent too long delays in email generation
-			$wgMemc->set( $this->createCacheKeyForMobileApplicationsLinks(),
-				$mobileApplicationsLinks, time() + self::MOBILE_APPLICATIONS_LINKS_EVICTION_TIME );
-		}
-
-		return $mobileApplicationsLinks;
-	}
-
-	private function createCacheKeyForMobileApplicationsLinks() {
-		return wfMemcKey( 'mobileApplicationsLinks' );
-	}
-
-	/**
-	 * @return string or boolean (because of Http::request)
-	 */
-	private function fetchMobileApplicationsDetails() {
-		// currently it does not matter if Android or iOS value is added, data is returned for both Android and iOS
-		return Http::request( 'GET',
-			'https://services.wikia.com/mobile-applications/platform/android' );
-	}
-
-	/**
-	 * @param string response - raw, not parsed response
-	 *
-	 * @return boolean
-	 */
-	private function applicationsExistIn( $response ) {
-		return strpos( $response, 'wikia_id":' . $this->wg->CityId ) !== false;
-	}
-
-	/**
-	 * @param array $mobileApplications - array containing information about mobile applications
-	 * @return array - containing mobile applications links
-	 */
-	private function traverseThrough( $mobileApplications ) {
-		$mobileApplicationsLinks = [];
-		$siteId = $this->wg->CityId;
-
-		foreach ( $mobileApplications['apps'] as $app ) {
-			foreach ( $app['languages'] as $language ) {
-				if ( $language['wikia_id'] == $siteId ) {
-					if ( $app['android_release'] ) {
-						$release = $app['android_release'];
-						$mobileApplicationsLinks[self::ANDROID_PLATFORM] =
-							"https://play.google.com/store/apps/details?id=$release";
-					}
-					if ( $app['ios_release'] ) {
-						$release = $app['ios_release'];
-						$mobileApplicationsLinks[self::IOS_PLATFORM] =
-							"https://itunes.apple.com/us/app/id$release";
-					}
-					break;
-				}
-			}
-
-			if ( !empty( $mobileApplicationsLinks ) ) {
-				break;
-			}
-		}
-
-		return $mobileApplicationsLinks;
 	}
 
 	/**
